댓글에 동영상 업로드 기능 및 업로드된 동영상 스트리밍 기능 구현

+++ src/main/webapp/WEB-INF/jsp/article/detail.jsp
@@ -1,12 +1,7 @@
-<%@ page import="java.util.List"%>
-<%@ page import="com.sbs.kig.at.dto.Article"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 	pageEncoding="UTF-8"%>
-
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
-
 <c:set var="pageTitle" value="게시물 상세내용" />
-
 <%@ include file="../part/head.jspf"%>
 
 <div class="table-box con">
@@ -17,7 +12,7 @@
 				<td>${article.id}</td>
 			</tr>
 			<tr>
-				<th>작성일</th>
+				<th>날짜</th>
 				<td>${article.regDate}</td>
 			</tr>
 			<tr>
@@ -32,11 +27,8 @@
 	</table>
 </div>
 
-
-<!-- Param을 사용할수있는건 넘겨준 쿼리스트링 뿐이구나 -->
 <c:if test="${isLogined}">
 	<h2 class="con">댓글 작성</h2>
-
 	<script>
 		function ArticleWriteReplyForm__submit(form) {
 			form.body.value = form.body.value.trim();
@@ -46,9 +38,15 @@
 				return;
 			}
 			var startUploadFiles = function(onSuccess) {
+				if(form.file__reply__0__common__attachment__1.value.length == 0 && form.file__reply__0__common__attachment__1.value == null) {
+					onSuccess();
+					return;
+				}
+				
 				var fileUploadFormData = new FormData(form);
 				fileUploadFormData.delete("relTypeCode");
 				fileUploadFormData.delete("relId");
+				
 				$.ajax({
 					url : './../file/doUploadAjax',
 					data : fileUploadFormData,
@@ -59,7 +57,7 @@
 					success : onSuccess
 				});
 			}
-			alert('이제 fileIds(' + fileIdsStr + ')를 doWriteReplyAjax에서 처리해야 한다.');
+			
 			var startWriteReply = function(fileIdsStr, onSuccess) {
 				$.ajax({
 					url : './../reply/doWriteReplyAjax',
@@ -76,13 +74,24 @@
 			};
 			startUploadFiles(function(data) {
 				var idsStr = data.body.fileIdsStr;
+	
+				if(data && data.body && data.body.fileIdStr) {
+					idStr = data.body.fileIdsStr;
+				}
+				
 				startWriteReply(idsStr, function(data) {
+					if(data.msg) {
+						alert(data.msg);
+					}
+					
 					form.body.value = '';
+					form.file__reply__0__common__attachment__1.value = '';
+					ArticleWriteReplyForm__submitDone = false;
 				});
 			});
 		}
 	</script>
-
+	
 	<form class="table-box con form1"
 		onsubmit="ArticleWriteReplyForm__submit(this); return false;">
 		<input type="hidden" name="relTypeCode" value="article" /> 
@@ -99,11 +108,11 @@
 					</td>
 				</tr>
 				<tr>
-					<th>첨부1</th>
+					<th>첨부1 비디오</th>
 					<td>
 						<div class="form-control-box">
 							<input type="file" accept="video/*" capture
-								name="file__articleReply__0__common__attachment__1">
+								name="file__reply__0__common__attachment__1">
 						</div>
 					</td>
 				</tr>
@@ -114,19 +123,17 @@
 			</tbody>
 		</table>
 	</form>
-
 </c:if>
 
 <h2 class="con">댓글 리스트</h2>
-
 <div class="reply-list-box table-box con">
 	<table>
 		<colgroup>
-			<col width="80" />
-			<col width="180" />
-			<col width="180" />
-			<col />
-			<col width="200" />
+			<col width="80">
+			<col width="180">
+			<col width="180">
+			<col>
+			<col width="200">
 		</colgroup>
 		<thead>
 			<tr>
@@ -138,7 +145,6 @@
 			</tr>
 		</thead>
 		<tbody>
-
 		</tbody>
 	</table>
 </div>
@@ -153,6 +159,7 @@
 	background-color: rgba(0, 0, 0, 0.4);
 	display: none;
 }
+
 .reply-modify-form-modal-actived .reply-modify-form-modal {
 	display: flex;
 }
@@ -180,10 +187,8 @@
 
 <script>
 	var ReplyList__$box = $('.reply-list-box');
-	var ReplyList__$tbody = ReplyList__$box.find('tbody')
-
+	var ReplyList__$tbody = ReplyList__$box.find('tbody');
 	var ReplyList__lastLodedId = 0;
-
 	var ReplyList__submitModifyFormDone = false;
 
 	function ReplyList__submitModifyForm(form) {
@@ -191,7 +196,6 @@
 			alert('처리중입니다.');
 			return;
 		}
-		
 		form.body.value = form.body.value.trim();
 		if (form.body.value.length == 0) {
 			alert('내용을 입력해주세요.');
@@ -200,7 +204,6 @@
 		}
 		var id = form.id.value;
 		var body = form.body.value;
-		
 		ReplyList__submitModifyFormDone = true;
 		$.post('../reply/doModifyReplyAjax', {
 			id : id,
@@ -216,13 +219,12 @@
 			ReplyList__submitModifyFormDone = false;
 		}, 'json');
 	}
+	
 	function ReplyList__showModifyFormModal(el) {
 		$('html').addClass('reply-modify-form-modal-actived');
 		var $tr = $(el).closest('tr');
 		var originBody = $tr.data('data-originBody');
-		
 		var id = $tr.attr('data-id');
-
 		var form = $('.reply-modify-form-modal form').get(0);
 		form.id.value = id;
 		form.body.value = originBody;
@@ -234,16 +236,16 @@
 	
 	function ReplyList__loadMoreCallback(data) {
 		if (data.body.replies && data.body.replies.length > 0) {
-			ReplyList__lastLodedId  = data.body.replies[data.body.replies.length - 1].id;
+			ReplyList__lastLodedId = data.body.replies[data.body.replies.length - 1].id;
 			ReplyList__drawReplies(data.body.replies);
 		}
 		setTimeout(ReplyList__loadMore, 2000);
 	}
-
+	
 	function ReplyList__loadMore() {
 		$.get('../reply/getForPrintReplies', {
 			articleId : param.id,
-			from : ReplyList__lastLodedId  + 1
+			from : ReplyList__lastLodedId + 1
 		}, ReplyList__loadMoreCallback, 'json');
 	}
 	
@@ -255,15 +257,12 @@
 	}
 	
 	function ReplyList__delete(el) {
-		if ( confirm('삭제 하시겠습니까?') == false ) {
+		if (confirm('삭제 하시겠습니까?') == false) {
 			return;
 		}
-		
 		var $tr = $(el).closest('tr');
-		
 		var id = $tr.attr('data-id');
-
-		$.post('../reply/doDeleteReplyAjax', {
+		$.post('./../reply/doDeleteReplyAjax', {
 			id : id
 		}, function(data) {
 			$tr.remove();
@@ -272,28 +271,33 @@
 	
 	function ReplyList__drawReply(reply) {
 		var html = '';
-		
 		html += '<tr data-id="' + reply.id + '">';
 		html += '<td>' + reply.id + '</td>';
 		html += '<td>' + reply.regDate + '</td>';
 		html += '<td>' + reply.extra.writer + '</td>';
-		html += '<td class="reply-body">' + reply.body + '</td>';
+		html += '<td>';
+		html += '<div class="reply-body">' + reply.body + '</div>';
+		if (reply.extra.file__common__attachment__1) {
+            var file = reply.extra.file__common__attachment__1;
+            html += '<video controls src="http://localhost:8085/usr/file/streamVideo?id=' + file.id + '">video not supported</video>';
+        }
+		
+		html += '</td>';
 		html += '<td>';
 		if (reply.extra.actorCanDelete) {
 			html += '<button type="button" onclick="ReplyList__delete(this);">삭제</button>';
 		}
+		
 		if (reply.extra.actorCanModify) {
 			html += '<button type="button" onclick="ReplyList__showModifyFormModal(this);">수정</button>';
 		}
+		
 		html += '</td>';
 		html += '</tr>';
-
 		var $tr = $(html);
 		$tr.data('data-originBody', reply.body);
 		ReplyList__$tbody.prepend($tr);
 	}
-	
 	ReplyList__loadMore();
 </script>
-
 <%@ include file="../part/foot.jspf"%>
